JavaScript/solution_1: Remove unnecessary pinned down versions (#854)

+++ PrimeV/solution_1/primes.v
@@ -67,7 +67,7 @@ fn (sieve Sieve) print_results(show_results bool, duration time.Duration, passes
 
 	avg := f64(duration / passes)
 	count_primes := sieve.count_primes()
-	valid := (count_primes == dictionary[sieve.sieve_size.str()])
+	valid := (count_primes == u64(dictionary[sieve.sieve_size.str()]))
 	eprintln('Passes: $passes, Time: $duration, Avg: $avg, Limit: $sieve.sieve_size, Count1: $count, Count2: $count_primes, Valid: $valid')
 
 	println('marghidanu;$passes;$duration;1;algorithm=base,faithful=yes')
